% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate}
\alias{generate}
\title{Generate a list of authors and their affiliations from a dramatis.personae YAML file.}
\usage{
generate(file = "dramatis.personae.yaml")
}
\arguments{
\item{file}{The file path.}
}
\value{
A list containing a list of authors and a list of affiliations.
}
\description{
Generate a list of authors and their affiliations from a dramatis.personae YAML file.
}
