% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Create a dramatis.personae YAML file template.}
\usage{
initialize(file = "dramatis.personae.yaml")
}
\arguments{
\item{file}{The file path.}
}
\description{
Create a dramatis.personae YAML file template.
}
